Customer Registration

+++ backend/src/main/java/com/netcracker/edu/backend/repository/CustomerEntityRepository.java
@@ -3,5 +3,5 @@ package com.netcracker.edu.backend.repository;
 import com.netcracker.edu.backend.entity.CustomerEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CustomerEntityRepository extends CrudRepository<CustomerEntity, Integer> {
+public interface CustomerEntityRepository extends CrudRepository<CustomerEntity, Long> {
 }
